change 8
Trying to fix the 'TypeError: object of type 'function' has no len()' error because of the draw_path function In order to fix this, I need to make sure that it is the path variable that is passed to draw_path, not the function itself. I should be using the dijkstra_path and a_star_path variables, which contain the shortest paths from A to S. Now I fixed the problem.